huffman passes 33 tests of 36

+++ lab-10/lab-10/code.h
@@ -6,7 +6,7 @@
 
 typedef struct _code
 {
-    int code;
+    long long code;
     int length;
 } HuffmanCode;
